+++ meerschaum/core/Pipe/_fetch.py
@@ -7,7 +7,9 @@ Functions for fetching new data into the Pipe
 """
 
 from __future__ import annotations
+from datetime import timedelta
 from meerschaum.utils.typing import Optional, Any
+from meerschaum.config import get_config
 
 def fetch(
         self,
@@ -98,3 +100,30 @@ def fetch(
         connector_plugin.deactivate_venv(debug=debug)
 
     return df
+
+
+def get_backtrack_interval(self, debug: bool = False) -> Union[timedelta, int]:
+    """
+    Get the chunk interval to use for this pipe.
+
+    Returns
+    -------
+    The backtrack interval (`timedelta` or `int`) to use with this pipe's `datetime` axis.
+    """
+    default_backtrack_minutes = get_config('pipes', 'parameters', 'fetch', 'backtrack_minutes')
+    configured_backtrack_minutes = self.parameters.get('fetch', {}).get('backtrack_minutes', None)
+    backtrack_minutes = (
+        configured_backtrack_minutes
+        if configured_backtrack_minutes is not None
+        else default_backtrack_minutes
+    )
+
+    dt_col = self.columns.get('datetime', None)
+    if dt_col is None:
+        return timedelta(minutes=backtrack_minutes)
+
+    dt_dtype = self.dtypes.get(dt_col, 'datetime64[ns]')
+    if 'datetime' in dt_dtype.lower():
+        return timedelta(minutes=backtrack_minutes)
+
+    return backtrack_minutes